Back from holiday!
Finally back from a 2 week holiday and can start developing again!

Added a different type of enemy thats rare to find and when killed gives you HP.

Also added a placeholder boss. Still need to implement attacks so that its actually a threat but thats pretty much fully implemented.

When the boss dies I need to give the player a choice of what to upgrade. (Make the timescale 0 and add UI elements that when clicked upgrade the players stats)